Backend/django/requirements.txt
asgiref==3.5.0
async-generator==1.10
attrs==21.4.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
cryptography==36.0.1
Django==4.0.2
djangorestframework==3.13.1
EasyProcess==1.1
gunicorn==20.1.0
h11==0.13.0
idna==3.3
mariadb==1.0.9
mysqlclient==2.1.0
outcome==1.1.0
pycparser==2.21
pyOpenSSL==22.0.0
pytz==2021.3
PyVirtualDisplay==2.2
requests==2.27.1
selenium==4.1.0
sniffio==1.2.0
sortedcontainers==2.4.0
sqlparse==0.4.2
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
wsproto==1.0.0
xlrd==2.0.1
